Home page component
This covers generating component using the cli and covering the
associated files changed. Then moving the files around and playing
with those same files again. The home page is then added to the
app.component.html file.

Some data is defined in the home.componenet.ts file and displayed
in home.component.html.