Added more pages and a router
Added an about page and a contact page. The contact page has a form
that is bound to an object in contact.component.ts that is also
bound to the form using two way data binding. The FormsModule was
also added to the app.component.module file.

The router can load all three pages and has a catch all for anything
else. The RouterModule was also added to the app.component.module
file. The <router-outlet> component was also added to the html file
for the main app.